Fix merge fallout in tests and docs
- gc integration tests: expired_removed widened to u64 upstream; align
  the Case table field type
- rename slash-command tests: drop billing_surface_visible (fork removed
  the consumer billing surface)
- inline auto-compact flow tests: call handle_set_session_model through
  Arc with the new skip_prompt_rewrite parameter
- docs/internal: add 25-enterprise.md + 22-environment-variables.md
  mirrors of the [features] registry (upstream's
  registered_features_are_documented tripwire referenced files that were
  never committed)

+++ crates/codegen/xai-fast-worktree/src/api/gc/integration_tests.rs
@@ -951,7 +951,7 @@ fn per_kind_age_expiry_reclaims_listed_kinds_and_keeps_the_rest() {
         records: Vec<(&'static str, WorktreeKind, i64)>,
         opts: gc::GcOptions,
         survivors: Vec<&'static str>,
-        expired_removed: usize,
+        expired_removed: u64,
     }
     let cases = vec![
         Case {
